events: moved fast event emitter to own file
